requirements.txt
streamlit==1.41.1
google-genai==0.3.0
vertexai==1.71.1
pillow==11.1.0
python-dotenv==1.0.1